x64, x86 separate support
/ Fixed Pointer.ReadPointer cast bug
/ Fixed exportresolver EDT address conditional typo
* Now all project can target x64 or x86 independently

Currently my MemoryModule is suffering on AccessViolationError. I will compare op-to-op with DLLFromMemory.cs and fix it.

+++ StealthModule/Pointer.cs
@@ -40,7 +40,7 @@ namespace StealthModule
 
         public T Read<T>() => (T)Marshal.PtrToStructure(this, typeof(T));
 
-        public Pointer ReadPointer() => (Pointer)Marshal.PtrToStructure(this, typeof(IntPtr));
+        public Pointer ReadPointer() => (Pointer)(IntPtr)Marshal.PtrToStructure(this, typeof(IntPtr));
 
         public void Write<T>(T buffer) => Marshal.StructureToPtr(buffer, this, false);
